# docs/swagger/openapi.yaml
openapi: "3.0.3"
info:
  title: Workflow Execution Service
  version: "1.0.0"
  description: >
    Runtime control plane for the BPMN-driven Workflow Engine. Manages workflow instances,
    task projections for dashboard queries, and user signals (claim, complete, defer, reassign).
    All identity and tenant context is injected by the Envoy API Gateway — callers never set
    these headers directly.
servers:
  - url: https://api.internal/execution
    description: Internal API gateway base (service prefix only; paths include /api/v1/)
security:
  - GatewayTenantContext: []
    UserContext: []
...
components:
  securitySchemes:
    GatewayTenantContext:
      type: apiKey
      in: header
      name: x-tenant-id
      description: Tenant UUID injected by the API Gateway.
    UserContext:
      type: apiKey
      in: header
      name: x-user-id
      description: Authenticated user UUID injected by the API Gateway.
    AdminRole:
      type: apiKey
      in: header
      name: x-tenant-roles
      description: >
        Comma-separated tenant roles injected by the API Gateway. Admin-only endpoints
        require at least one of: tenant_admin, tenant_owner.
    InternalServiceAuth:
      type: apiKey
      in: header
      name: x-internal-token
      description: >
        Service-to-service credential for the internal route group (RequireInternalToken,
        INTERNAL_API_TOKEN) — the same convention Definition Service already ships. These
        routes are NOT exposed through the Envoy public gateway and do NOT carry gateway
        identity headers; they are restricted to in-mesh callers by a Kubernetes
        NetworkPolicy AND authenticated by this shared secret — defense-in-depth, not either
        control alone (§9.2). The tenant context for these routes is taken from the request
        payload, not from x-tenant-id.
  parameters:
    TenantIdHeader:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant UUID (injected by gateway).
    UserIdHeader:
      name: x-user-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Authenticated user UUID (injected by gateway).
    TenantRolesHeader:
      name: x-tenant-roles
      in: header
      required: true
      schema:
        type: string
      description: >
        Comma-separated tenant roles. Must include tenant_admin or tenant_owner
        for admin-only endpoints.
    DepartmentsHeader:
      name: x-departments
      in: header
      required: false
      schema:
        type: string
      description: Comma-separated department UUIDs the caller belongs to (gateway-injected).
...
    TaskIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Workflow task UUID.
...
    CursorQuery:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: >
        Opaque keyset cursor encoding the previous page's last (created_at|occurred_at, id)
        pair (§5.9). Omit for the first page. An unparseable or tampered cursor returns 400
        BAD_REQUEST rather than silently restarting.
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: >
        Page size. Values above 100 are clamped, not rejected; values <= 0 are rejected
        (§5.9).
...
  schemas:
...
    TaskStatus:
      type: string
      enum: [READY, IN_PROGRESS, COMPLETED, DEFERRED, FAILED, SUPERSEDED]
      description: >
        READY: no assignment has been claimed. IN_PROGRESS: at least one assignment claimed,
        not all completed. COMPLETED: all active assignments completed (Temporal advances).
        DEFERRED: reviewer sent task back; a regression task was created.
        FAILED: swept up in an instance-cancel/-terminate/FAILED cascade — never an
        independent per-task failure. SUPERSEDED: bypassed by an admin instance-force-forward.
...
    TaskSummary:
      type: object
      required:
        - id
        - workflow_instance_id
        - tenant_id
        - node_key
        - task_type
        - department_id
        - status
        - record_version
        - assignee_mode
        - assignee_count
        - created_at
      properties:
        id:
          type: string
          format: uuid
        workflow_instance_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        node_key:
          type: string
          description: BPMN node ID matching the compiled plan.
          example: "review_stage_1"
        task_type:
          type: string
          example: "userTask"
        department_id:
          type: string
          format: uuid
          description: Snapshotted from the compiled plan at task creation.
        status:
          $ref: "#/components/schemas/TaskStatus"
        record_version:
          type: integer
          description: >
            Optimistic lock counter. Incremented on every Claim or assignment change.
            Clients must echo the record_version they last observed in mutation requests.
          example: 2
        assignee_mode:
          $ref: "#/components/schemas/AssigneeMode"
        assignee_count:
          type: integer
          description: Expected number of active assignments for this task.
          example: 2
        due_at:
          type: string
          format: date-time
          nullable: true
          description: SLA deadline snapshotted from the compiled plan at task creation.
        deferred_from_task_id:
          type: string
          format: uuid
          nullable: true
          description: If this task is a regression created by a Defer signal, the ID of the parent task.
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    TaskDetail:
      allOf:
        - $ref: "#/components/schemas/TaskSummary"
        - type: object
          properties:
            payload_json:
              type: object
              description: >
                Runtime context passed from the instance start request at task creation.
                Read-only; matches the context_json provided when the instance was started.
              additionalProperties: true
            assignments:
              type: array
              items:
                $ref: "#/components/schemas/AssignmentRecord"
              description: All assignment records for this task (active and historical).
...
  responses:
    BadRequest:
      description: Malformed JSON body or missing required fields.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/bad-request"
            title: "Bad Request"
            status: 400
            detail: "field 'version' is required"
            instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/claim"
            code: "BAD_REQUEST"
    Unauthorized:
      description: Missing or invalid x-tenant-id / x-user-id gateway headers.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/unauthorized"
            title: "Unauthorized"
            status: 401
            detail: "x-tenant-id header is required"
            instance: "/api/v1/instances"
            code: "UNAUTHORIZED"
...
    InternalError:
      description: Unhandled server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/internal-error"
            title: "Internal Server Error"
            status: 500
            detail: "an unexpected error occurred"
            instance: "/api/v1/instances"
            code: "INTERNAL_ERROR"
...
paths:
...
  /api/v1/tasks:
    get:
      operationId: listTasks
      tags: [Tasks]
      summary: List workflow tasks
      description: >
        Returns a paginated list of tasks for the caller's tenant. Supports filtering by
        status, assignee, department, instance, and due date. Used by task dashboards.
        Additionally scoped to the same assignee/department/admin visibility rule as
        GET /tasks/:id (§9.2) — non-visible tasks are simply absent from the page, not an error.
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/CursorQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
          description: Filter by task status.
        - name: assignee_user_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter to tasks with an active assignment for this user.
        - name: department_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter to tasks belonging to this department.
        - name: instance_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter to tasks belonging to this workflow instance.
        - name: due_before
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Include only tasks with due_at before this timestamp.
      responses:
        "200":
          description: Keyset page of workflow tasks (§5.9).
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskSummary"
                  next_cursor:
                    type: string
                    nullable: true
                    description: Opaque cursor for the next page; omitted/null on the final page.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tasks/{id}:
    get:
      operationId: getTask
      tags: [Tasks]
      summary: Get task detail
      description: >
        Returns task detail including payload_json and all assignment records (active and
        historical). Visible only if the caller is (or was) an assignee, the caller's
        x-departments includes the task's department_id, or the caller holds
        tenant_admin/tenant_owner (§9.2). Returns 403 NOT_AUTHORIZED_FOR_RESOURCE otherwise —
        not 404, since the task is already confirmed to exist in the caller's own tenant.
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantRolesHeader"
        - $ref: "#/components/parameters/DepartmentsHeader"
        - $ref: "#/components/parameters/TaskIdPath"
      responses:
        "200":
          description: Task detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: >
            NOT_AUTHORIZED_FOR_RESOURCE — caller is not an assignee, not in the task's
            department, and not an admin.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://errors.bcbp.io/execution/not-authorized-for-resource"
                title: "Not Authorized For Resource"
                status: 403
                detail: "caller is not the assignee, department member, or admin for this task"
                instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6"
                code: "NOT_AUTHORIZED_FOR_RESOURCE"
        "404":
          $ref: "#/components/responses/TaskNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tasks/{id}/claim:
    post:
      operationId: claimTask
      tags: [Tasks]
      summary: Claim a task assignment
      description: >
        The calling user claims their active unclaimed assignment for this task.
        Sets claimed_at on the assignment row and increments task.record_version.
        Sends a task-claim signal to the Temporal workflow.
        Returns 403 if the user has no active assignment.
        Returns 409 TASK_ALREADY_CLAIMED if the user's assignment is already claimed.
        Returns 409 RECORD_VERSION_CONFLICT if the supplied record_version is stale.
        Returns 409 CLAIM_NOT_APPLICABLE if the task's assignee_mode is single — claim exists
        only to establish a lead among multiple assignees (assignee_mode=all).
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TaskIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [record_version]
              properties:
                record_version:
                  type: integer
                  minimum: 0
                  description: >
                    The task record_version the client last observed. Must match the current
                    task.record_version; mismatch returns RECORD_VERSION_CONFLICT.
                  example: 0
      responses:
        "202":
          description: Claim accepted; returns the task with its incremented record_version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Caller has no active assignment for this task.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://errors.bcbp.io/execution/assignee-not-assigned"
                title: "Not Assigned"
                status: 403
                detail: "caller does not have an active assignment for this task"
                instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/claim"
                code: "NOT_ASSIGNEE"
        "404":
          $ref: "#/components/responses/TaskNotFound"
        "409":
          description: >
            State conflict. See code field:
            TASK_ALREADY_CLAIMED — the caller's assignment is already claimed;
            RECORD_VERSION_CONFLICT — the supplied version is stale (another claim beat you);
            INVALID_TASK_STATE — the task is not in READY or IN_PROGRESS status;
            CLAIM_NOT_APPLICABLE — the task's assignee_mode is single, so claim does not apply.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                AlreadyClaimed:
                  value:
                    type: "https://errors.bcbp.io/execution/task-already-claimed"
                    title: "Task Already Claimed"
                    status: 409
                    detail: "this assignment has already been claimed"
                    instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/claim"
                    code: "TASK_ALREADY_CLAIMED"
                ClaimNotApplicable:
                  value:
                    type: "https://errors.bcbp.io/execution/claim-not-applicable"
                    title: "Claim Not Applicable"
                    status: 409
                    detail: "claim only applies to tasks with assignee_mode=all"
                    instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/claim"
                    code: "CLAIM_NOT_APPLICABLE"
                ConcurrencyViolation:
                  value:
                    type: "https://errors.bcbp.io/execution/task-concurrency-violation"
                    title: "Task Concurrency Violation"
                    status: 409
                    detail: "task record_version mismatch; refresh and retry"
                    instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/claim"
                    code: "RECORD_VERSION_CONFLICT"
                IdempotencyReplay:
                  value:
                    type: "https://errors.bcbp.io/execution/idempotency-key-replay"
                    title: "Idempotency Key Replay"
                    status: 409
                    detail: "Idempotency-Key reused with a different request body"
                    instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/claim"
                    code: "IDEMPOTENCY_KEY_REPLAY"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [Tasks]
      summary: Complete a task assignment
      description: >
        The calling user submits their completion payload for their active assignment.
        Sets completed_at and result_json on the assignment. If all active assignments for
        the task are now complete, the task status transitions to COMPLETED and a
        task-complete signal is sent to Temporal (which then advances the workflow).
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TaskIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [record_version]
              properties:
                record_version:
                  type: integer
                  minimum: 0
                  description: The task record_version the client last observed.
                  example: 1
                result_json:
                  type: object
                  additionalProperties: true
                  description: >
                    Completion payload from this assignee. Optional — not bound as
                    required by the handler. Schema is workflow-defined; the service
                    stores it opaquely. Included in the TaskCompleted event.
                  example:
                    approved: true
                    notes: "All documents are in order."
      responses:
        "202":
          description: >
            Completion recorded; returns the task as of this write. The response does not
            carry a separate all_done flag — check the returned status field instead (rev 1.23).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Caller has no active claimed assignment for this task.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://errors.bcbp.io/execution/assignee-not-assigned"
                title: "Not Assigned"
                status: 403
                detail: "caller does not have an active claimed assignment for this task"
                instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/complete"
                code: "NOT_ASSIGNEE"
        "404":
          $ref: "#/components/responses/TaskNotFound"
        "409":
          description: >
            State conflict. See code field:
            RECORD_VERSION_CONFLICT — version mismatch;
            INVALID_TASK_STATE — task is not IN_PROGRESS;
            IDEMPOTENCY_KEY_REPLAY — same key reused with a different request body.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                IdempotencyReplay:
                  value:
                    type: "https://errors.bcbp.io/execution/idempotency-key-replay"
                    title: "Idempotency Key Replay"
                    status: 409
                    detail: "Idempotency-Key reused with a different request body"
                    instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/complete"
                    code: "IDEMPOTENCY_KEY_REPLAY"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/tasks/{id}/defer:
    post:
      operationId: deferTask
      tags: [Tasks]
      summary: Defer a task (send it back)
      description: >
        The calling user defers their claimed assignment — indicating the task cannot
        proceed as-is (e.g. missing documents, incomplete prior step). The task status
        transitions to DEFERRED, a regression task is created at the previous step,
        and a task-defer signal is sent to Temporal. The response returns the deferred
        task itself (rev 1.23) — it does not currently carry the new regression task's ID;
        a caller that needs it must look it up separately (accepted gap, not yet designed).
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TaskIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [record_version]
              properties:
                record_version:
                  type: integer
                  minimum: 0
                  description: The task record_version the client last observed.
                  example: 1
                reason:
                  type: string
                  maxLength: 500
                  description: >
                    Human-readable reason for deferral. Optional — not bound as required
                    by the handler. Stored on the TaskDeferred event and on the
                    regression task's assignment reason field.
                  example: "Supporting documents are missing from the tender package."
      responses:
        "202":
          description: Defer accepted; returns the deferred task (now in DEFERRED status).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Caller has no active claimed assignment for this task.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://errors.bcbp.io/execution/assignee-not-assigned"
                title: "Not Assigned"
                status: 403
                detail: "caller does not have an active claimed assignment for this task"
                instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/defer"
                code: "NOT_ASSIGNEE"
        "404":
          $ref: "#/components/responses/TaskNotFound"
        "409":
          description: >
            RECORD_VERSION_CONFLICT — version mismatch;
            INVALID_TASK_STATE — task is not IN_PROGRESS or has no previous step to defer to;
            IDEMPOTENCY_KEY_REPLAY — same key reused with a different request body.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                IdempotencyReplay:
                  value:
                    type: "https://errors.bcbp.io/execution/idempotency-key-replay"
                    title: "Idempotency Key Replay"
                    status: 409
                    detail: "Idempotency-Key reused with a different request body"
                    instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/defer"
                    code: "IDEMPOTENCY_KEY_REPLAY"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...